add a sample script to validate Metric Hub definitions
I used a script similar to this to validate definition changes in https://github.com/mozilla/metric-hub/pull/483 and https://github.com/mozilla/metric-hub/pull/523 after merging. @jmsilverman suggested it could be helpful to check the script into the repo.

Open to a discussion about whether or not this is useful and, if not, if there's something we can do instead.