Fix klar active_users data_sources
The `active_users_view` for Klar products currently has the filter `is_mobile`. However, Klar is not included as a mobile KPI product, so `is_mobile` will always be `false`.

This PR changes the filter to "Klar iOS" or "Klar Android", which will filter only on valid sources of Klar client_ids (excluding BrowserStack, for example).